1.0.2 a45
live data broken out into own file
