+++ autots/datasets/__init__.py
@@ -2,9 +2,9 @@
 Tools for Importing Sample Data
 """
 
+from autots.datasets._live import load_live_daily
 from autots.datasets._base import (
     load_daily,
-    load_live_daily,
     load_monthly,
     load_yearly,
     load_hourly,
